feat(Entity,AdapterEntity): Added the notion of `properties` and `attributes`
Renamed `IRawEntityAttributes` & `IRawAdapterEntityAttributes` to `IEntityAttributes` & `IEntityProperties.
Added distinction between properties (that can contain the managed properties `id` & `idHash`, and that are related strictly to an adapter) and attributes, which contains the business data.
Removed the `toObject()` method
Moved raw query-related interfaces to the `QueryLanguage.Raw` namespace.
`DataSourceQuerier` is now generic with 2 parameters, that inputs an object  and outputs sonething inheriting the first.