chore(LICENSE): Changed license from GPLv3 to Apache2.0
GPLv3 imposed the using softwares to be under the same license, which can be very restrictive for users. The Apache2.0 is more compliant on using softwares, and impose the users to document changes made on Diaspora. Those changes may be then used in Diaspora itself, if the fork license allows us to do so.